Add MUI example
